Changes to enable peregrine
