+++ gen3-integration-tests/conftest.py
@@ -61,8 +61,6 @@ def pytest_configure(config):
         setup.create_program_project()
         # Generate test data
         setup.generate_graph_data()
-        # Restarting indexd service for Sheepdog and Peregrine
-        # assert kube_setup_service(pytest.namespace, "indexd")
 
     # Compute root url for portal
     try:
