--- gen3-integration-tests/gen3_ci/jenkins_jobs/mutate-manifest-for-guppy-test.groovy
@@ -1,62 +1,0 @@
-/*
-  String parameter NAMESPACE
-    e.g., qa-anvil
-
-  Archived artifacts -
-*/
-pipeline {
-    agent {
-        node {
-            label 'gen3-ci-worker'
-        }
-    }
-    stages {
-        stage('Clean old workspace') {
-            steps {
-                cleanWs()
-            }
-        }
-        stage('Initial Setup') {
-            steps {
-                // cloud-automation
-                checkout([
-                  $class: 'GitSCM',
-                  branches: [[name: 'refs/heads/master']],
-                  doGenerateSubmoduleConfigurations: false,
-                  extensions: [[$class: 'RelativeTargetDirectory', relativeTargetDir: 'cloud-automation']],
-                  submoduleCfg: [],
-                  userRemoteConfigs: [[credentialsId: 'PlanXCyborgUserJenkins', url: 'https://github.com/uc-cdis/cloud-automation.git']]
-                ])
-            }
-        }
-        stage('Mutate Manifest for Guppy Test') {
-            options {
-                    timeout(time: 5, unit: 'MINUTES')   // timeout on this stage
-                }
-            steps {
-                dir("guppy-gen3-setup"){
-                    script {
-                        try {
-                            sh '''#!/bin/bash +x
-                            set -e
-                            export GEN3_HOME=\$WORKSPACE/cloud-automation
-                            export KUBECTL_NAMESPACE=\${NAMESPACE}
-                            source $GEN3_HOME/gen3/gen3setup.sh
-
-                            # Mutate guppy config to point guppy to pre-defined Canine ETL'ed data
-                            gen3 mutate-guppy-config-for-guppy-test
-
-                            # Validate pods to roll up
-                            gen3 kube-wait4-pods || true
-                            '''
-                        }
-                        catch (ex) {
-                            metricsHelper.writeMetricWithResult(STAGE_NAME, false)
-                            pipelineHelper.handleError(e)
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
