+++ gen3-integration-tests/tests/test_data_upload.py
@@ -12,10 +12,7 @@ from cdislogging import get_logger
 from services.dataclient import DataClient
 from services.fence import Fence
 from services.indexd import Indexd
-from services.peregrine import Peregrine
-from services.coremetadata import CoreMetaData
-from pages.login import LoginPage
-from pages.core_metadata_page import CoreMetadataPage
+
 from services.graph import GraphDataTools
 from utils import data_upload
 
@@ -45,32 +42,23 @@ class FileNode:
 @pytest.mark.indexd
 @pytest.mark.sheepdog
 class TestDataUpload:
-    @classmethod
-    def setup_class(cls):
+    def setup_method(self, method):
+        # Delete local small file created in setup_method
         auth = Gen3Auth(refresh_token=pytest.api_keys["main_account"])
         sd_tools = GraphDataTools(
             auth=auth, program_name="jnkns", project_code="jenkins"
         )
         sd_tools.delete_nodes()
-
-    def setup_method(self, method):
         # Create a local small file to upload. Store its size and hash
         with open(file_path, "w") as file:
             file.write(file_content)
         # Create a local large file (size 7MB)
-        with open(file_path, "w") as file:
+        with open(big_file_path, "w") as file:
             file.write(file_content)
 
     def teardown_method(self, method):
-        # Delete local small file created in setup_method
         os.remove(file_path)
-        # Delete all existing nodes prior to running the test cases
-        logger.info("Deleting any existing nodes before test case execution")
-        auth = Gen3Auth(refresh_token=pytest.api_keys["main_account"])
-        sd_tools = GraphDataTools(
-            auth=auth, program_name="jnkns", project_code="jenkins"
-        )
-        sd_tools.delete_nodes()
+        os.remove(big_file_path)
 
     def test_file_upload_and_download_via_api(self):
         """
@@ -102,16 +90,21 @@ class TestDataUpload:
             presigned_url = fence_upload_res["url"]
 
             # Check blank record was created in indexd
-            file_node = FileNode(did=file_guid, props={"md5sum": file_md5, "file_size": file_size})
+            file_node = FileNode(
+                did=file_guid, props={"md5sum": file_md5, "file_size": file_size}
+            )
 
             indexd.get_record(indexd_guid=file_node.did)
 
             # fail to submit metadata for this file without hash and size
             try:
                 metadata_response = sd_tools.submit_graph_and_file_metadata(
-                    file_guid, file_size, file_md5)
+                    file_guid, file_size, file_md5
+                )
             except Exception as e:
-                assert "400" in f"{e}", f"Linking metadata to file without hash and size should not be possible.\n{metadata_response}"
+                assert (
+                    "400" in f"{e}"
+                ), f"Linking metadata to file without hash and size should not be possible.\n{metadata_response}"
 
             # check that we CANNOT download the file (there is no URL in indexd yet)
             signed_url_res = fence.createSignedUrl(
@@ -150,7 +143,7 @@ class TestDataUpload:
         finally:
             indexd.delete_files(created_guids)
 
-    '''def test_user_without_role_cannot_upload(self):
+    def test_user_without_role_cannot_upload(self):
         """
         Scenario: User without role can't upload
         Steps:
@@ -203,7 +196,9 @@ class TestDataUpload:
             # Upload the file to the S3 bucket using the presigned URL
             data_upload.upload_file_to_s3(presigned_url, file_path, file_size)
 
-            file_node = FileNode(did=file_guid, props={"md5sum": file_md5, "file_size": file_size})
+            file_node = FileNode(
+                did=file_guid, props={"md5sum": file_md5, "file_size": file_size}
+            )
 
             # wait for the indexd listener to add size, hashes and URL to the record
             data_upload.wait_upload_file_updated_from_indexd_listener(indexd, file_node)
@@ -225,9 +220,12 @@ class TestDataUpload:
             # no metadata linking after delete
             try:
                 metadata_response = sd_tools.submit_graph_and_file_metadata(
-                    file_guid, file_size, file_md5)
+                    file_guid, file_size, file_md5
+                )
             except Exception as e:
-                assert "400" in f"{e}", f"Linking metadata to file without hash and size should not be possible.\n{metadata_response}"
+                assert (
+                    "400" in f"{e}"
+                ), f"Linking metadata to file without hash and size should not be possible.\n{metadata_response}"
 
             # no download after delete
             fence.createSignedUrl(id=file_guid, user="main_account", expectedStatus=404)
@@ -269,7 +267,9 @@ class TestDataUpload:
             # Upload the file to the S3 bucket using the presigned URL
             data_upload.upload_file_to_s3(presigned_url, file_path, file_size)
 
-            file_node = FileNode(did=file_guid, props={"md5sum": file_md5, "file_size": file_size})
+            file_node = FileNode(
+                did=file_guid, props={"md5sum": file_md5, "file_size": file_size}
+            )
 
             # wait for the indexd listener to add size, hashes and URL to the record
             data_upload.wait_upload_file_updated_from_indexd_listener(indexd, file_node)
@@ -317,162 +317,3 @@ class TestDataUpload:
             fence.check_file_equals(signed_url_res, file_content)
         finally:
             indexd.delete_files(created_guids)
-
-    def test_core_metadata_page(self, page: LoginPage):
-        """
-        Scenario: Test core metadata page
-        Steps:
-            1. Create nodes using sheepdog. Verify nodes are present.
-            2. Create a file node
-            3. Get core metadata using node in step 2
-            4. Reteive object_id from metadata record recieved
-            5. Load metadata page using object_id and verify the elements
-            6. Delete file node and delete indexd record
-            7. Delete all nodes. Verify all nodes are deleted.
-        """
-        peregrine = Peregrine()
-        login_page = LoginPage()
-        core_metadata_page = CoreMetadataPage()
-        coremetadata = CoreMetaData()
-        auth = Gen3Auth(refresh_token=pytest.api_keys["main_account"])
-        sd_tools = GraphDataTools(
-            auth=auth, program_name="jnkns", project_code="jenkins"
-        )
-        nodes = sd_tools.get_path_to_file()
-        # Create nodes using sheepdog. Verify nodes are present.
-        for key, val in nodes.items():
-            sd_tools._submit_record(record=val)
-
-        node = sd_tools.get_file_node()
-        sd_tools._submit_record(record=node)
-        node.indexd_guid = sd_tools.get_did_from_file_id(guid=node.unique_id)
-        metadata = coremetadata.get_core_metadata(file=node, user="main_account")
-        peregrine.see_json_core_metadata(file=node, metadata=metadata)
-
-        # Perform login and logout operations using main_account to create a login record for audit service to access
-        logger.info("Logging in with mainAcct")
-        login_page.go_to(page)
-        login_page.login(page)
-        core_metadata_page.goto_metadata_page(page, metadata.json()["object_id"])
-        core_metadata_page.verify_metadata_page_elements(page)
-        login_page.logout(page)
-
-        sd_tools._delete_record(unique_id=node.unique_id)
-
-        # Delete all remaining records
-        sd_tools.delete_nodes()
-
-    def test_core_metadata(self):
-        """
-        Scenario: Test core metadata
-        Steps:
-            1. Create nodes and verify they are added using main_account
-            2. Add file node and verify it is added using main_account
-            3. Identify the endpoint of coremetadata using the peregrine version
-            4. Get metadata record for the GUID from file node using application/json format
-            5. Validate file_name, GUID, Type and data_format of file node matches in metadata
-            6. Get metadata record for the GUID from file node using x-bibtex format
-            7. Validate file_name, GUID, Type and data_format of file node matches in metadata
-            8. Delete all nodes
-        """
-        peregrine = Peregrine()
-        coremetadata = CoreMetaData()
-        auth = Gen3Auth(refresh_token=pytest.api_keys["main_account"])
-        sd_tools = GraphDataTools(
-            auth=auth, program_name="jnkns", project_code="jenkins"
-        )
-        nodes = sd_tools.get_path_to_file()
-        # Create nodes using sheepdog. Verify nodes are present.
-        for key, val in nodes.items():
-            sd_tools._submit_record(record=val)
-
-        node = sd_tools.get_file_node()
-        sd_tools._submit_record(record=node)
-        node.indexd_guid = sd_tools.get_did_from_file_id(guid=node.unique_id)
-        metadata = coremetadata.get_core_metadata(file=node, user="main_account")
-        peregrine.see_json_core_metadata(file=node, metadata=metadata)
-
-        metadata = coremetadata.get_core_metadata(
-            file=node, user="main_account", format="x-bibtex"
-        )
-        coremetadata.see_bibtex_core_metadata(file=node, metadata=metadata)
-
-        sd_tools._delete_record(unique_id=node.unique_id)
-
-        # Delete all remaining records
-        sd_tools.delete_nodes()
-
-    def test_core_metadata_invalid_object_id(self):
-        """
-        Scenario: Test core metadata invalid object_id
-        Steps:
-            1. Create nodes and verify they are added using main_account
-            2. Get invalid_file node details and update did with wrong value
-            3. Identify the endpoint of coremetadata using the peregrine version
-            4. Get metadata record for the GUID from file node using application/json format
-            5. Step 4 should fail with 404 error as object/did is not present
-            6. Delete all nodes
-        """
-        coremetadata = CoreMetaData()
-        auth = Gen3Auth(refresh_token=pytest.api_keys["main_account"])
-        sd_tools = GraphDataTools(
-            auth=auth, program_name="jnkns", project_code="jenkins"
-        )
-        nodes = sd_tools.get_path_to_file()
-        # Create nodes using sheepdog. Verify nodes are present.
-        for key, val in nodes.items():
-            sd_tools._submit_record(record=val)
-
-        invalid_node = sd_tools.get_file_node()
-        invalid_node.props["object_id"] = "invalid_object_id"
-        invalid_node.indexd_guid = "invalid_object_id"
-        metadata = coremetadata.get_core_metadata(
-            file=invalid_node, user="main_account", expected_status=404
-        )
-        coremetadata.see_core_metadata_error(
-            metadata=metadata, message='object_id "invalid_object_id" not found'
-        )
-
-        # Delete all remaining records
-        sd_tools.delete_nodes()
-
-    def test_core_metadata_no_permission(self):
-        """
-        Scenario: Test core metadata no permission
-        Steps:
-            1. Create nodes and verify they are added using main_account
-            2. Add file node and verify it is added using main_account
-            3. Identify the endpoint of coremetadata using the peregrine version
-            4. Get metadata record for the GUID from file node using application/json format and invalid authorization
-            5. Step 4 should fail with 401 error as invalid authorization was passed
-            6. Verify "Authentication Error: could not parse authorization header" message was recieved
-            7. Delete all nodes
-        """
-        coremetadata = CoreMetaData()
-        auth = Gen3Auth(refresh_token=pytest.api_keys["main_account"])
-        sd_tools = GraphDataTools(
-            auth=auth, program_name="jnkns", project_code="jenkins"
-        )
-        nodes = sd_tools.get_path_to_file()
-        # Create nodes using sheepdog. Verify nodes are present.
-        for key, val in nodes.items():
-            sd_tools._submit_record(record=val)
-
-        node = sd_tools.get_file_node()
-        sd_tools._submit_record(record=node)
-        node.indexd_guid = sd_tools.get_did_from_file_id(guid=node.unique_id)
-        metadata = coremetadata.get_core_metadata(
-            file=node,
-            user="main_account",
-            expected_status=401,
-            invalid_authorization=True,
-        )
-        coremetadata.see_core_metadata_error(
-            metadata=metadata,
-            message="Authentication Error: could not parse authorization header",
-        )
-
-        sd_tools._delete_record(unique_id=node.unique_id)
-
-        # Delete all remaining records
-        sd_tools.delete_nodes()'''
