+++ gen3-integration-tests/tests/test_guppy_service.py
@@ -5,7 +5,6 @@ import os
 import pytest
 
 from cdislogging import get_logger
-from utils.gen3_admin_tasks import mutate_manifest_for_guppy_test
 from services.guppy import Guppy
 
 
@@ -17,7 +16,6 @@ class TestGuppyService:
     @classmethod
     def setup_class(cls):
         guppy = Guppy()
-        assert mutate_manifest_for_guppy_test(pytest.namespace)
         assert guppy.validate_guppy_status("main_account", 200)
 
     def test_guppy_test_query_1(self):
