+++ gen3-integration-tests/tests/test_manifest_service.py
